+++ assets/javascripts/kitten/components/sliders/slider-tooltip.js
